feat: autorized user can see all functions

+++ project.py
@@ -21,7 +21,7 @@ APPLICATION_NAME = "Restaurant Menu Application"
 
 
 # Connect to Database and create database session
-engine = create_engine('sqlite:///restaurantmenu.db')
+engine = create_engine('sqlite:///restaurantmenuwithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -112,7 +112,7 @@ def gconnect():
 
     user_id = getUserID(login_session['email'])
     if not user_id:
-        user_id = create_engine(login_session)
+        user_id = createUser(login_session)
     login_session['user_id'] = user_id
 
 
@@ -188,7 +188,10 @@ def showRestaurants():
     DBSession = sessionmaker(bind=engine)
     session = DBSession()
     restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-    return render_template('restaurants.html', restaurants=restaurants)
+    if 'username' not in login_session:
+        return render_template('publicrestaurants.html', restaurants=restaurants)
+    else:
+        return render_template('restaurants.html', restaurants=restaurants)
 
 
 # Create a new restaurant
@@ -235,6 +238,8 @@ def deleteRestaurant(restaurant_id):
         return redirect('/login')
     restaurantToDelete = session.query(
         Restaurant).filter_by(id=restaurant_id).one()
+    if restaurantToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(restaurantToDelete)
         flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -251,9 +256,13 @@ def showMenu(restaurant_id):
     DBSession = sessionmaker(bind=engine)
     session = DBSession()
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    creator = getUserInfo(restaurant.user_id)
     items = session.query(MenuItem).filter_by(
         restaurant_id=restaurant_id).all()
-    return render_template('menu.html', items=items, restaurant=restaurant)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template('publicmenu.html', items=items, restaurant=restaurant, creator=creator)
+    else:
+        return render_template('menu.html', items=items, restaurant=restaurant, creator=creator)
 
 
 # Create a new menu item
@@ -320,6 +329,8 @@ def deleteMenuItem(restaurant_id, menu_id):
 
 
 def getUserID(email):
+    DBSession = sessionmaker(bind=engine)
+    session = DBSession()
     try:
         user = session.query(User).filter_by(email = email).one()
         return user.id
@@ -328,11 +339,15 @@ def getUserID(email):
 
 
 def getUserInfo(user_id):
+    DBSession = sessionmaker(bind=engine)
+    session = DBSession()
     user = session.query(User).filter_by(id = user_id).one()
     return user
 
 
 def createUser(login_session):
+    DBSession = sessionmaker(bind=engine)
+    session = DBSession()
     newUser = User(name = login_session['username'], email= login_session['email'], picture = login_session['picture'])
     session.add(newUser)
     session.commit()
